Providers

+++ Software Project/src/Etisalat.java
@@ -1,5 +1,5 @@
 
 public interface Etisalat extends Provider{
-	public int getAmount();
+	public double getAmount();
 
 }
