+++ Software Project/src/Orange.java
@@ -1,5 +1,5 @@
 
 public interface Orange extends Provider{
-	public int getAmount();
+	public double getAmount();
 
 }
