+++ Software Project/src/Vodafone.java
@@ -1,5 +1,5 @@
 
 public interface Vodafone extends Provider{
-	public int getAmount();
+	public double getAmount();
 
 }
