+++ Software Project/src/We.java
@@ -1,5 +1,5 @@
 
 public interface We extends Provider{
-	public int getAmount();
+	public double getAmount();
 
 }
